Move code into archives package and split run and message specific code into their own modules
