+++ archives/archives_test.go
@@ -1,4 +1,4 @@
-package archiver
+package archives
 
 import (
 	"compress/gzip"
@@ -16,7 +16,7 @@ import (
 )
 
 func setup(t *testing.T) *sqlx.DB {
-	testDB, err := ioutil.ReadFile("testdb.sql")
+	testDB, err := ioutil.ReadFile("../testdb.sql")
 	assert.NoError(t, err)
 
 	db, err := sqlx.Open("postgres", "postgres://temba:temba@localhost:5432/archiver_test?sslmode=disable&TimeZone=UTC")
@@ -487,7 +487,7 @@ func TestArchiveOrgRuns(t *testing.T) {
 		assert.Equal(t, time.Date(2017, 10, 10, 0, 0, 0, 0, time.UTC), created[11].StartDate)
 		assert.Equal(t, DayPeriod, created[11].Period)
 		assert.Equal(t, 1, created[11].RecordCount)
-		assert.Equal(t, int64(427), created[11].Size)
+		assert.Equal(t, int64(428), created[11].Size)
 		assert.Equal(t, "bf08041cef314492fee2910357ec4189", created[11].Hash)
 
 		assert.Equal(t, 12, len(deleted))
