+++ archives/s3.go
@@ -1,4 +1,4 @@
-package archiver
+package archives
 
 import (
 	"context"
